// frontend/src/backend.d.ts
export interface CabBooking {
    status: BookingStatus;
    bookingId: string;
    dropoffLocation: string;
    cabType: string;
    bookingDate: bigint;
    guest: Principal;
    pickupTime: bigint;
    totalPrice: bigint;
    pickupLocation: string;
}
export interface HotelBooking {
    status: BookingStatus;
    depositAmount: bigint;
    bookingId: string;
    remainingAmount: bigint;
    checkInDate: bigint;
    bookingDate: bigint;
    guest: Principal;
    paymentPlan: PaymentPlan;
    checkOutDate: bigint;
    totalPrice: bigint;
    roomType: string;
    depositPercentage: bigint;
}
export interface PaymentRecord {
    status: Variant_new_verified_rejected;
    customer: Principal;
    note: string;
    reference: string;
    paymentId: string;
    timestamp: bigint;
    amount: bigint;
}
export interface SalesOrder {
    status: BookingStatus;
    deliveryAddress: string;
    remainingAmount: bigint;
    customer: Principal;
    orderDate: bigint;
    email: string;
    orderId: string;
    upfrontAmount: bigint;
    paymentPlan: PaymentPlan;
    contactNumber: string;
    products: Array<Product>;
    totalPrice: bigint;
}
export interface CabAvailability {
    availableCount: bigint;
    cabType: string;
}
export interface RoomType {
    features: Array<string>;
    imageUrls: Array<string>;
    pricePerNight: bigint;
    name: string;
    offerPrice: bigint;
}
export interface CabType {
    name: string;
    pricePerTrip: bigint;
    imageUrl: string;
    offerPrice: bigint;
    capacity: bigint;
}
export interface DateRange {
    checkIn: bigint;
    checkOut: bigint;
}
export interface ActingDriverRequest {
    status: BookingStatus;
    serviceDate: bigint;
    vehicleType: string;
    requestId: string;
    serviceDetails: string;
    bookingDate: bigint;
    guest: Principal;
}
export interface Inquiry {
    status: Variant_new_closed_reviewed;
    contactInfo: string;
    customer: Principal;
    message: string;
    timestamp: bigint;
    inquiryId: string;
}
export interface Product {
    imageUrls: Array<string>;
    name: string;
    description: string;
    productId: string;
    offerPrice: bigint;
    category: string;
    price: bigint;
}
export interface UserProfile {
    name: string;
    email: string;
    phone: string;
}
export enum BookingStatus {
    cancelled = "cancelled",
    pending = "pending",
    confirmed = "confirmed"
}
export enum PaymentPlan {
    cod = "cod",
    partialUpfront = "partialUpfront",
    depositRequired = "depositRequired",
    fullUpfront = "fullUpfront"
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export enum Variant_new_closed_reviewed {
    new_ = "new",
    closed = "closed",
    reviewed = "reviewed"
}
export enum Variant_new_verified_rejected {
    new_ = "new",
    verified = "verified",
    rejected = "rejected"
}